use of the automatic formater black

+++ spacewalks/eva_data_analysis.py
@@ -1,19 +1,20 @@
 import matplotlib.pyplot as plt
 import pandas as pd
 
+
 def read_json_to_dataframe(input_file):
-    print(f'Reading JSON file {input_file}')
+    print(f"Reading JSON file {input_file}")
     # Read the data from a JSON file into a Pandas dataframe
-    eva_df = pd.read_json(input_file, convert_dates=['date'])
-    eva_df['eva'] = eva_df['eva'].astype(float)
+    eva_df = pd.read_json(input_file, convert_dates=["date"])
+    eva_df["eva"] = eva_df["eva"].astype(float)
     # Clean the data by removing any incomplete rows and sort by date
     eva_df.dropna(axis=0, inplace=True)
-    eva_df.sort_values('date', inplace=True)
+    eva_df.sort_values("date", inplace=True)
     return eva_df
 
 
 def write_dataframe_to_csv(df, output_file):
-    print(f'Saving to CSV file {output_file}')
+    print(f"Saving to CSV file {output_file}")
     # Save dataframe to CSV file for later analysis
     df.to_csv(output_file, index=False)
 
@@ -22,9 +23,9 @@ def write_dataframe_to_csv(df, output_file):
 
 print("--START--")
 
-input_file = open('eva-data.json', 'r')
-output_file = open('eva-data.csv', 'w')
-graph_file = 'cumulative_eva_graph.png'
+input_file = open("eva-data.json", "r")
+output_file = open("eva-data.csv", "w")
+graph_file = "cumulative_eva_graph.png"
 
 # Read the data from JSON file
 eva_data = read_json_to_dataframe(input_file)
@@ -32,13 +33,15 @@ eva_data = read_json_to_dataframe(input_file)
 # Convert and export data to CSV file
 write_dataframe_to_csv(eva_data, output_file)
 
-print(f'Plotting cumulative spacewalk duration and saving to {graph_file}')
+print(f"Plotting cumulative spacewalk duration and saving to {graph_file}")
 # Plot cumulative time spent in space over years
-eva_data['duration_hours'] = eva_data['duration'].str.split(":").apply(lambda x: int(x[0]) + int(x[1])/60)
-eva_data['cumulative_time'] = eva_data['duration_hours'].cumsum()
-plt.plot(eva_data['date'], eva_data['cumulative_time'], 'ko-')
-plt.xlabel('Year')
-plt.ylabel('Total time spent in space to date (hours)')
+eva_data["duration_hours"] = (
+    eva_data["duration"].str.split(":").apply(lambda x: int(x[0]) + int(x[1]) / 60)
+)
+eva_data["cumulative_time"] = eva_data["duration_hours"].cumsum()
+plt.plot(eva_data["date"], eva_data["cumulative_time"], "ko-")
+plt.xlabel("Year")
+plt.ylabel("Total time spent in space to date (hours)")
 plt.tight_layout()
 plt.savefig(graph_file)
 plt.show()
